requirements.txt
brotlipy==0.7.0
cchardet==2.1.7
importlib-metadata==6.6.0
jupyter==1.0.0
jupyterlab==3.5.3
langchain==0.0.171
markdown-it-py==2.2.0
pinecone-client==2.2.1
pip-chill==1.0.3
pyopenssl==23.0.0
pypdf==3.8.1
pyqt5-sip==12.11.0
pysocks==1.7.1
python-dotenv==1.0.0
selenium==4.9.1
sip==6.6.2
tabulate==0.9.0
unstructured==0.7.1
streamlit==1.25.0